Clean up the local build
