--- jupyterlab/src/builder.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-export * from '@jupyterlab/extension-builder';
