--- jupyterlab/src/package.json
@@ -1,48 +1,0 @@
-{
-  "private": true,
-  "name": "@jupyterlab/application-top",
-  "scripts": {
-    "build": "webpack"
-  },
-  "dependencies": {
-    "@jupyterlab/about-extension": "^0.2.0",
-    "@jupyterlab/application-extension": "^0.2.0",
-    "@jupyterlab/apputils-extension": "^0.2.0",
-    "@jupyterlab/codemirror-extension": "^0.2.0",
-    "@jupyterlab/completer-extension": "^0.2.0",
-    "@jupyterlab/console-extension": "^0.2.0",
-    "@jupyterlab/csvwidget-extension": "^0.2.0",
-    "@jupyterlab/docmanager-extension": "^0.2.0",
-    "@jupyterlab/docregistry-extension": "^0.2.0",
-    "@jupyterlab/editorwidget-extension": "^0.2.0",
-    "@jupyterlab/faq-extension": "^0.2.0",
-    "@jupyterlab/filebrowser-extension": "^0.2.0",
-    "@jupyterlab/help-extension": "^0.2.0",
-    "@jupyterlab/imagewidget-extension": "^0.2.0",
-    "@jupyterlab/inspector-extension": "^0.2.0",
-    "@jupyterlab/landing-extension": "^0.2.0",
-    "@jupyterlab/launcher-extension": "^0.2.0",
-    "@jupyterlab/markdownwidget-extension": "^0.2.0",
-    "@jupyterlab/notebook-extension": "^0.2.0",
-    "@jupyterlab/rendermime-extension": "^0.2.0",
-    "@jupyterlab/running-extension": "^0.2.0",
-    "@jupyterlab/services-extension": "^0.2.0",
-    "@jupyterlab/shortcuts-extension": "^0.2.0",
-    "@jupyterlab/terminal-extension": "^0.2.0",
-    "@jupyterlab/tooltip-extension": "^0.2.0"
-  },
-  "devDependencies": {
-    "@jupyterlab/application": "^0.2.0",
-    "@jupyterlab/default-theme": "^0.2.0",
-    "css-loader": "^0.27.3",
-    "es6-promise": "^4.1.0",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
-    "font-awesome": "^4.7.0",
-    "fs-extra": "^2.1.2",
-    "json-loader": "^0.5.4",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^2.2.1"
-  }
-}
