--- jupyterlab/src/typings.d.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-
-/// <reference path="../../packages/codemirror/typings/codemirror/codemirror.d.ts"/>
-/// <reference path="../../packages/coreutils/typings/path-posix/path-posix.d.ts"/>
-/// <reference path="../../packages/coreutils/typings/url-parse/url-parse.d.ts"/>
-/// <reference path="../../packages/extension-builder/typings/webpack-config/webpack-config.d.ts"/>
-/// <reference path="../../packages/rendermime/typings/ansi_up/ansi_up.d.ts"/>
-/// <reference path="../../packages/terminal/typings/xterm/xterm.d.ts"/>
